+++ src/main/java/com/DTOLayer/DTOEntity/eventDTO/RequestEventDTO.java
@@ -1,4 +1,4 @@
-package com.DTOLayer.DTOEntity;
+package com.DTOLayer.DTOEntity.eventDTO;
 
 import java.time.LocalDateTime;
